+++ laserchicken/test_feature_extractor/test_feature_map.py
@@ -11,11 +11,12 @@ class FeatureMapTests(unittest.TestCase):
         expected_features = ['point_density', 'echo_ratio', 'eigenv_1', 'eigenv_2', 'eigenv_3', 'normal_vector_1',
                              'normal_vector_2', 'normal_vector_3', 'slope', 'entropy_z', 'pulse_penetration_ratio',
                              'sigma_z', 'median_z', 'max_z', 'min_z',
-                             'range_z', 'var_z', 'mean_z', 'std_z', 'coeff_var_z', 'skew_z', 'kurto_z', 'skew_norm_z',
-                             'mean_norm_z', 'std_norm_z', 'coeff_var_norm_z', 'var_norm_z', 'max_norm_z', 'min_norm_z',
-                             'range_norm_z', 'kurto_norm_z', 'entropy_norm_z', 'median_norm_z',
-                             'density_absolute_mean_z', 'density_absolute_mean_norm_z', 'perc_15_z',
-                             'perc_99_normalized_height']
+                             'range_z', 'var_z', 'mean_z', 'std_z', 'coeff_var_z', 'skew_z', 'kurto_z', 'skew_normalized_height',
+                             'mean_normalized_height','std_normalized_height', 'coeff_var_normalized_height', 'var_normalized_height', 'min_normalized_height',
+                             'max_normalized_height', 'range_normalized_height', 'kurto_normalized_height', 'entropy_normalized_height',
+                             'median_normalized_height', 'max_intensity',
+                             'density_absolute_mean_z', 'density_absolute_mean_normalized_height', 'perc_15_z',
+                             'perc_99_normalized_height', 'max_intensity', 'min_intensity', 'range_intensity', 'mean_intensity', 'std_intensity', 'coeff_var_intensity']
         for feature in expected_features:
             self.assertIn(feature, feature_extractor.FEATURES)